backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
psycopg2-binary==2.9.6
python-dotenv==1.0.1
httpx==0.25.0
PyJWT==2.8.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
